refactor: Add new crate iroh-node-util (name TBD)...
and move the node and net proto, client and server there
